re-organize the facet related file
