+++ packages/ui/src/features/facets/index.ts
@@ -1,14 +1,16 @@
+export * from "./utils/SearchInterface";
+
+export { StringFacet } from "./utils/StringFacet";
+export { StringListFacet } from "./utils/StringListFacet";
+export { TypeFacet } from "./utils/TypeFacet";
+export { VStringFacet } from "./utils/VStringFacet";
+export { VTypeFacet } from "./utils/VTypeFacet";
+export { VUserFacet } from "./utils/VUserFacet";
+
 export { EnvironmentFacet } from "./EnvironmentFacet";
-export * from "./FacetsNav";
-export { StringFacet } from "./StringFacet";
-export { StringListFacet } from "./StringListFacet";
-export { TypeFacet } from "./TypeFacet";
-
-export * from "./VFacetsNav";
-export { VStringFacet } from "./VStringFacet";
-export { VTypeFacet } from "./VTypeFacet";
-export { VUserFacet } from "./VUserFacet";
-
 export { RunsFacetsNav, useRunsFilterGroups, useRunsFilterHandler } from "./RunsFacetsNav";
 export { WorkflowExecutionsFacetsNav, useWorkflowExecutionsFilterGroups, useWorkflowExecutionsFilterHandler } from "./WorkflowExecutionsFacetsNav";
 export { DocumentsFacetsNav, useDocumentFilterGroups, useDocumentFilterHandler } from "./DocumentsFacetsNav";
+export { InteractionsFacetsNav, useInteractionsFilterGroups, useInteractionsFilterHandler } from "./InteractionsFacetsNav";
+export { PromptsFacetsNav, usePromptsFilterGroups, usePromptsFilterHandler } from "./PromptsFacetsNav";
+export { CollectionsFacetsNav, useCollectionsFilterGroups, useCollectionsFilterHandler } from "./CollectionsFacetsNav";
