--- packages/ui/src/features/facets/utils.tsx
@@ -1,9 +1,0 @@
-import { FacetBucket, FacetNameBucket } from '@vertesia/common';
-
-export function facetOptionLabel(bucket: FacetBucket) {
-    return `${bucket._id} (${bucket.count})`;
-}
-
-export function facetOptionNameLabel(bucket: FacetNameBucket) {
-    return `${bucket.name} (${bucket.count})`;
-}
